move middleware and hostoptions
