+++ cmd/appconnector-rp/main.go
@@ -15,9 +15,9 @@ import (
 	"syscall"
 
 	"github.com/go-logr/logr"
+	"github.com/project-radius/radius/pkg/armrpc/hostoptions"
 	"github.com/project-radius/radius/pkg/connectorrp/frontend"
 	"github.com/project-radius/radius/pkg/corerp/dataprovider"
-	"github.com/project-radius/radius/pkg/corerp/hostoptions"
 	"github.com/project-radius/radius/pkg/radlogger"
 	"github.com/project-radius/radius/pkg/telemetry/metrics/metricsservice"
 	mh "github.com/project-radius/radius/pkg/telemetry/metrics/metricsservice/hostoptions"
